Fix rendering of areas using a shader (Texture).
If an opaque colour is set that takes precedence,
and by default the colour is set to be white.
So no colour is set explicitly set a fully
transparent one instead.
This affects things like nature reserves, military
areas, graveyards, ...
It is most obvious along the coast of southern Sweden
which has quite a few military shooting ranges
(observable in all but the LightRS renderers).